fix power

+++ packages/runtime/src/operators/power.ts
@@ -1,8 +1,8 @@
-import {Integer} from "../types";
+import {Float} from "../types";
 import {ICharacter} from "../types/_character";
 import {INumeric} from "../types/_numeric";
 import {parse} from "./_parse";
 
 export function power(left: INumeric | ICharacter | string | number, right: INumeric | ICharacter | string | number) {
-  return new Integer().set(Math.pow(parse(left), parse(right)));
+  return new Float().set(Math.pow(parse(left), parse(right)));
 }
